+++ frontend/src/layout/index.jsx
@@ -1,129 +1,32 @@
-import { useState } from 'react';
-import {
-    createStyles,
-    Header,
-    Container,
-    Group,
-    Burger,
-    Paper,
-    Transition,
-    rem,
-} from '@mantine/core';
-import { useDisclosure } from '@mantine/hooks';
-import ThemeToggle from './components/ThemeToggle';
-import Logo from './components/Logo';
+import { Outlet } from 'react-router-dom';
+import Footer from './components/Footer';
+import Header from './components/Header';
 
-const HEADER_HEIGHT = rem(60);
-
-const useStyles = createStyles((theme) => ({
-    root: {
-        position: 'relative',
-        zIndex: 1,
-    },
-
-    dropdown: {
-        position: 'absolute',
-        top: HEADER_HEIGHT,
-        left: 0,
-        right: 0,
-        zIndex: 0,
-        borderTopRightRadius: 0,
-        borderTopLeftRadius: 0,
-        borderTopWidth: 0,
-        overflow: 'hidden',
-
-        [theme.fn.largerThan('sm')]: {
-            display: 'none',
+export default function () {
+    const links = [
+        {
+            "link": "/about",
+            "label": "About"
         },
-    },
-
-    header: {
-        display: 'flex',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        height: '100%',
-    },
-
-    links: {
-        [theme.fn.smallerThan('sm')]: {
-            display: 'none',
+        {
+            "link": "/campgrounds",
+            "label": "Campgrounds"
         },
-    },
-
-    burger: {
-        [theme.fn.largerThan('sm')]: {
-            display: 'none',
+        {
+            "link": "/Login",
+            "label": "Login"
         },
-    },
-
-    link: {
-        display: 'block',
-        lineHeight: 1,
-        padding: `${rem(8)} ${rem(12)}`,
-        borderRadius: theme.radius.sm,
-        textDecoration: 'none',
-        color: theme.colorScheme === 'dark' ? theme.colors.dark[0] : theme.colors.gray[7],
-        fontSize: theme.fontSizes.sm,
-        fontWeight: 500,
-
-        '&:hover': {
-            backgroundColor: theme.colorScheme === 'dark' ? theme.colors.dark[6] : theme.colors.gray[0],
+        {
+            "link": "/register",
+            "label": "Sign up"
         },
-
-        [theme.fn.smallerThan('sm')]: {
-            borderRadius: 0,
-            padding: theme.spacing.md,
-        },
-    },
-
-    linkActive: {
-        '&, &:hover': {
-            backgroundColor: theme.fn.variant({ variant: 'light', color: theme.primaryColor }).background,
-            color: theme.fn.variant({ variant: 'light', color: theme.primaryColor }).color,
-        },
-    },
-}));
-
-
-export default function ({ links }) {
-    const [opened, { toggle, close }] = useDisclosure(false);
-    const [active, setActive] = useState(links[0].link);
-    const { classes, cx } = useStyles();
-
-    const items = links.map((link) => (
-        <a
-            key={link.label}
-            href={link.link}
-            className={cx(classes.link, { [classes.linkActive]: active === link.link })}
-            onClick={(event) => {
-                event.preventDefault();
-                setActive(link.link);
-                close();
-            }}
-        >
-            {link.label}
-        </a>
-    ));
+    ]
 
     return (
-        <Header height={HEADER_HEIGHT} mb={120} className={classes.root}>
-            <Container className={classes.header}>
-                <Logo />
-                <Group spacing={5} className={classes.links}>
-                    {items}
-                    <ThemeToggle />
-                </Group>
-
-                <Burger opened={opened} onClick={toggle} className={classes.burger} size="sm" />
-
-                <Transition transition="pop-top-right" duration={200} mounted={opened}>
-                    {(styles) => (
-                        <Paper className={classes.dropdown} withBorder style={styles}>
-                            {items}
-                        </Paper>
-                    )}
-                </Transition>
-            </Container>
-        </Header>
+        <>
+            <Header links={links} />
+            <Outlet />
+            <Footer links={links} />
+        </>
     );
 }
